Hol long -> int 변경, 필요없는 코드 제거

+++ src/main/java/com/pos/repository/HolRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HolRepository extends JpaRepository<Hol, Long> {
+public interface HolRepository extends JpaRepository<Hol, Integer> {
 
 }
